Explisit ToJSON and FromJSON instances
While the existance of the `type` keword in haskell prevents easy
naming of fields, this does not prevent other JSON consumers from
using the undecorated names.

In this commit we write explisit To and From JSON class instances
to make the JSON spec more readable and not leak internal
implimentation details.